done with printf

+++ ft_putpointer.c
@@ -10,5 +10,5 @@
 void	ft_putpointer(void *p, int *count)
 {
 	ft_putstr("0x", count);
-	ft_puthex((unsigned long long)p, count);
+	ft_puthex((unsigned long)p, count);
 }
